Change the returnvalue of main()
In the original example from TI the main function has a void return value. This
is not C99 standard and cause trouble with the compiler.
I also changed the free passed arguments value to void, because main() is silent
about passing arguments and a main(void) says that main doesn't have to accept
any argument.