Get OOB-project to compile and link
In the previous commit the linker didn't work because there was no startup file
defined. This is fixed with this commit. I also added some other needed files to
build objects.

The OOB-example uses FreeRTOS and this must be defined by the compiler.
It was also necessary to add includes for FreeRTOS.

Some wrong paths at the includes were fixed too.