OOB example compiles and is running on device
To get the example running it was needed to add some preprocessor
ifdef in the main.c for the init of the vectortables.

It was also necessary to add some defines and fix some wrong defines
caused by copy-paste.
The -Dgcc define is needed by many examples and should be defined, because
if it is missing the code will compile without errors but won't run.